Tony/framework upgrade (#22)
* Build the basic project

* Build TWCore.Collections

* build fixes

* build fixes

* fixes

* Build fixes

* Build fixes

* fixes

* Improving build

* build fixes

* Fixes

* fix pipeline

* Fixes

* fixes

* fixes

* Changes

* fixes

* fixes

* Changes

* changes

* Changes

* Upload artifacts

* Improvements and fixes

* changes

* Improvements

* changes

* Update dependencies

* Changes

* Fixes

* Changes

* fix

* Fixes

* Changes

+++ src/TWCore.Serialization.NSerializer/SerializersTable.cs
@@ -1411,7 +1411,7 @@ namespace TWCore.Serialization.NSerializer
         protected void WriteSByte(sbyte value)
         {
             Span<byte> buffer = stackalloc byte[1];
-            BitConverter.TryWriteBytes(buffer, value);
+            BitConverter.TryWriteBytes(buffer, (short)value);
             Stream.Write(buffer);
         }
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
